o erro principal da matriz de perspectiv na verdade era um erro do renderer e foi corrigido. Matriz de perspectiva já calcula a deformacao do mesh mas inverte o eixo y ao passar pelo plano yz. Corrigir

+++ src/Renderer.java
@@ -24,21 +24,21 @@ class Renderer extends JPanel {
       DoubleMatrix faces = meshes[i].getFaces();
       for(int j=0; j<faces.columns; j++){
         DoubleMatrix t = faces.getColumn(j);
-        int p0w = (int)vertices.get(3,(int)t.get(0));
-        int p1w = (int)vertices.get(3,(int)t.get(1));
-        int p2w = (int)vertices.get(3,(int)t.get(2));
-        int p0x = X_OFFSET + (int)vertices.get(0,(int)t.get(0))*p0w;
-        int p0y = Y_OFFSET + (int)vertices.get(1,(int)t.get(0))*p0w;
-        int p0z = (int)vertices.get(2,(int)t.get(0))*p0w;
-        int p1x = X_OFFSET + (int)vertices.get(0,(int)t.get(1))*p1w;
-        int p1y = Y_OFFSET + (int)vertices.get(1,(int)t.get(1))*p1w;
-        int p1z = (int)vertices.get(2,(int)t.get(1))*p1w;
-        int p2x = X_OFFSET + (int)vertices.get(0,(int)t.get(2))*p2w;
-        int p2y = Y_OFFSET + (int)vertices.get(1,(int)t.get(2))*p2w;
-        int p2z = (int)vertices.get(2,(int)t.get(2))*p2w;
-        g.drawLine(p0x, p0y, p1x, p1y);
-        g.drawLine(p1x, p1y, p2x, p2y);
-        g.drawLine(p2x, p2y, p0x, p0y);
+        double p0w = vertices.get(3,(int)t.get(0));
+        double p1w = vertices.get(3,(int)t.get(1));
+        double p2w = vertices.get(3,(int)t.get(2));
+        double p0x = X_OFFSET + vertices.get(0,(int)t.get(0))*p0w;
+        double p0y = Y_OFFSET + vertices.get(1,(int)t.get(0))*p0w;
+        double p0z = vertices.get(2,(int)t.get(0))*p0w;
+        double p1x = X_OFFSET + vertices.get(0,(int)t.get(1))*p1w;
+        double p1y = Y_OFFSET + vertices.get(1,(int)t.get(1))*p1w;
+        double p1z = vertices.get(2,(int)t.get(1))*p1w;
+        double p2x = X_OFFSET + vertices.get(0,(int)t.get(2))*p2w;
+        double p2y = Y_OFFSET + vertices.get(1,(int)t.get(2))*p2w;
+        double p2z = vertices.get(2,(int)t.get(2))*p2w;
+        g.drawLine((int)p0x, (int)p0y, (int)p1x, (int)p1y);
+        g.drawLine((int)p1x, (int)p1y, (int)p2x, (int)p2y);
+        g.drawLine((int)p2x, (int)p2y, (int)p0x, (int)p0y);
         //g.drawLine((int)p2.getX(), (int)p2.getY(), (int)p0.getX(), (int)p0.getY());
         /*System.out.print("Desenhando triangulo " + j + " do mesh " + i + ":\n");
         System.out.print(
